+++ include/scopetree.h
@@ -20,12 +20,6 @@ typedef struct function {
 	arraylist_t *params;
 } function_t;
 
-typedef struct define {
-	char   *name;
-	char   *value;
-	type_t  type; //Needs to be identified
-} define_t;
-
 typedef struct scope {
 	struct scope *parent;
 	linkedlist_t *child;
@@ -36,21 +30,17 @@ typedef struct scope {
 	int to_line;
 } scope_t;
 
-
-define_t   *define_init(char *name, char *value);
 function_t *function_init(char *name, char *type, unsigned char type_is_pointer, arraylist_t *params);
 field_t    *field_init(char *name, char *type, unsigned char type_is_pointer);
 scope_t    *scope_init(scope_t *parent);
 
 field_t     *get_field_from_string(char *field);
-define_t    *get_define_from_string(char *define);
 arraylist_t *get_function_params(char *function_head);
 
 scope_t     *parse_scope(arraylist_t *file, unsigned int start_line, scope_t *parent_scope);
 
 char type_equals(type_t *type1, type_t *type2);
 
-void define_free(define_t *define);
 void field_free(field_t *field);
 void function_free(function_t *function);
 void scope_free(scope_t *scope);
